refactoring tests to accelerate CI validation
