--- test/unit/HashLists/ownership.Test.js
@@ -1,38 +1,0 @@
-/* eslint-env mocha */
-/* eslint-disable no-console */
-/* global artifacts, contract, describe, it, beforeEach */
-const chai = require('chai')
-const { assert } = chai
-const chaiAsPromised = require('chai-as-promised')
-chai.use(chaiAsPromised)
-
-const HashListLibrary = artifacts.require('HashListLibrary')
-const HashLists = artifacts.require('HashLists')
-
-contract('HashList', (accounts) => {
-    let hashListLibrary
-    let hashList
-    const owner = accounts[0]
-
-    beforeEach(async () => {
-        if (!hashList) {
-            hashListLibrary = await HashListLibrary.new()
-            HashLists.link('HashListLibrary', hashListLibrary.address)
-            hashList = await HashLists.new()
-            await hashList.initialize(owner, { from: owner })
-            const newAccountHash = await hashList.hash(accounts[1])
-            await hashList.methods['add(bytes32)'](newAccountHash, { from: owner })
-        }
-    })
-
-    describe('ownedBy', () => {
-        it('should return list owner', async () => {
-            assert.strictEqual(
-                await hashList.ownedBy(
-                    await hashList.hash(owner)
-                ),
-                owner
-            )
-        })
-    })
-})
